refactor: extract abstract test fixture to add tests with variations
